requirements.txt
boto3
typer
pyyaml
rich
setuptools
